chore: rename screenmanager.py to screen_manager.py
Renamed the file `screenmanager.py` to `screen_manager.py` to improve consistency with naming conventions. This change aligns file names with the snake_case style used throughout the project.
